chore: layer audit remediation — CLAUDE sync, PHPStan, L4 codified context boundary
- Align CLAUDE.md Layer 0/4/5 package lists with bin/check-package-layers
- Extend phpstan.neon (bimaaji, ai-observability, L6 PHP packages); fix strict-rules across touched packages
- Decouple CodifiedContextController from Telescope via Api session store interface + telescope adapter; telescope requires waaseyaa/api
- Exempt foundation Kernel/ and Http/Router/ from L0 static-use audit noise per CLAUDE
- Add segment_l2_covers_by_package.php; refresh L1/L3 segment script hints; PHPDoc @covers for codified context tests
- Regenerate build/layer0-audit and build/layer4-audit artifacts

Refs: #1339 #1340 #1342 #1344 #827 #848
Made-with: Cursor

+++ src/Analysis/AnomalyDetector.php
@@ -133,7 +133,7 @@ final class AnomalyDetector
     {
         $total = 0.0;
         foreach ($spans as $s) {
-            $total += (float) ($s['cost_usd'] ?? 0.0);
+            $total += $s['cost_usd'] ?? 0.0;
         }
 
         return $total;
